imagetotext/requirements.txt
torch~=1.13.1
requests~=2.28.2
transformers~=4.26.0
Pillow~=9.4.0
clip~=1.0
Flask~=2.2.2